requirements.txt
flask==2.0.3
pymongo==4.1.1
python-dotenv==0.20.0
gunicorn==20.1.0 